Refactor ObjectTransformer to TOAssembler and DOAssembler
