+++ backend/src/main/java/com/moodify/backend/api/controllers/UsersController.java
@@ -12,12 +12,12 @@ import org.springframework.web.server.ResponseStatusException;
 @RequestMapping("/users")
 public class UsersController {
     private final DatabaseService POSTGRES_SERVICE;
-    private final ObjectTransformer OBJECT_TRANSFORMER;
+    private final TOAssembler TO_OBJECT_ASSEMBLER;
 
     @Autowired
-    public UsersController(PostgresService USER_SERVICE, ObjectTransformer OBJECT_TRANSFORMER) {
+    public UsersController(PostgresService USER_SERVICE, TOAssembler TO_OBJECT_ASSEMBLER) {
         this.POSTGRES_SERVICE = USER_SERVICE;
-        this.OBJECT_TRANSFORMER = OBJECT_TRANSFORMER;
+        this.TO_OBJECT_ASSEMBLER = TO_OBJECT_ASSEMBLER;
     }
 
     @GetMapping({"findUserById/{userId}"})
@@ -25,7 +25,7 @@ public class UsersController {
     public UserTO findUserById(@PathVariable("userId") long userId) {
 
         try {
-            return this.OBJECT_TRANSFORMER.generateUserTOFrom(this.POSTGRES_SERVICE.findUserById(userId));
+            return this.TO_OBJECT_ASSEMBLER.generateUserTOFrom(this.POSTGRES_SERVICE.findUserById(userId));
         } catch (Exception e) {
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, e.getMessage(), e);
         }
